ref(helm): convert repo commands to new format
issue #1203
